Added basic O2 Meter

+++ Terminus/Assets/Scripts/Events/UpdateO2DisplayEvent.cs
@@ -6,6 +6,6 @@ using UnityEngine.Events;
 /// <summary>
 /// Event invoked to update oxygen UI gauge
 /// </summary>
-public class UpdateO2DisplayEvent : UnityEvent<int>
+public class UpdateO2DisplayEvent : UnityEvent<float>
 { 
 }
