+++ Terminus/Assets/Scripts/Managers/EventManager.cs
@@ -12,19 +12,19 @@ public static class EventManager
 
     // declare lists to hold invokers and listeners to update O2 display event
     static List<OxygenControl> updateO2Invokers = new List<OxygenControl>();
-    static List<UnityAction<int>> updateO2Listeners = new List<UnityAction<int>>();
+    static List<UnityAction<float>> updateO2Listeners = new List<UnityAction<float>>();
 
     // Adds given oxygen controller as invoker of update O2 event
     public static void AddUpdateO2Invoker(OxygenControl invoker)
     {
         // adds invoker to list and adds all listeners to this invoker
         updateO2Invokers.Add(invoker);
-        foreach (UnityAction<int> listener in updateO2Listeners)
+        foreach (UnityAction<float> listener in updateO2Listeners)
             invoker.AddUpdateO2Listener(listener);
     }
 
     // Adds given method as listener to update O2 event
-    public static void AddUpdateO2Listener(UnityAction<int> listener)
+    public static void AddUpdateO2Listener(UnityAction<float> listener)
     {
         // adds listener to list and to all invokers of event
         updateO2Listeners.Add(listener);
